# openapi.yaml
openapi: "3.0.3"
info:
  title: Secure Multi-Tier Platform API
  description: |
    Enterprise-grade SaaS API backend demonstrating multi-tier AWS architecture.
    Provides product catalogue management, order processing, user authentication
    via AWS Cognito (OAuth2 with PKCE), and platform health monitoring.
  version: "1.0.0"
  contact:
    name: Platform Team
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Local development (Docker Compose)
  - url: https://api.example.com
    description: Demo deployment (API Gateway + CloudFront)
...
security:
  - BearerAuth: []
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      summary: Register a new user
      description: |
        Creates a new user in the Cognito user pool with email as the username.
        Requires email verification. In local development, the user is auto-confirmed.
      tags:
        - auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            example:
              email: "user@example.com"
              password: "SecureP@ss123!"
              full_name: "Jane Smith"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
              example:
                message: "User registered successfully"
                user_sub: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                email: "user@example.com"
        "409":
          description: User with this email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: "USER_EXISTS"
                  message: "A user with this email already exists"
        "422":
          description: Validation error (invalid password or parameters)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      summary: Authenticate user and obtain tokens
      description: |
        Authenticates a user via Cognito and returns a JWT access token and
        refresh token. The access token should be included as a Bearer token
        in the Authorization header for subsequent API requests.
      tags:
        - auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              email: "user@example.com"
              password: "SecureP@ss123!"
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
              example:
                access_token: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
                refresh_token: "eyJjdHkiOiJKV1QiLCJlbmMiOiJBMjU2R0NNIi..."
                token_type: "Bearer"
                expires_in: 3600
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: "INVALID_CREDENTIALS"
                  message: "Invalid email or password"
        "403":
          description: User email not confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: "USER_NOT_CONFIRMED"
                  message: "User email has not been confirmed"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from AWS Cognito via OAuth2 Authorization Code
        flow with PKCE. Include the token in the Authorization header as:
        `Authorization: Bearer <token>`
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - full_name
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          description: User email address
        password:
          type: string
          minLength: 12
          maxLength: 128
          description: Password (min 12 chars, must include uppercase, lowercase, number, symbol)
        full_name:
          type: string
          minLength: 1
          maxLength: 255
          description: User's full name
    RegisterResponse:
      type: object
      required:
        - message
        - user_sub
        - email
      properties:
        message:
          type: string
          description: Success message
        user_sub:
          type: string
          description: Cognito user subject identifier (UUID)
        email:
          type: string
          format: email
          description: Registered email address
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          description: User email address
        password:
          type: string
          minLength: 1
          maxLength: 128
          description: User password
    LoginResponse:
      type: object
      required:
        - access_token
        - refresh_token
        - token_type
        - expires_in
      properties:
        access_token:
          type: string
          description: JWT access token for API authorization
        refresh_token:
          type: string
          description: Refresh token for obtaining new access tokens
        token_type:
          type: string
          enum:
            - Bearer
          default: Bearer
          description: Token type (always Bearer)
        expires_in:
          type: integer
          description: Token expiry time in seconds
          example: 3600

    # --- Products ---
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: "INVALID_CREDENTIALS"
            message:
              type: string
              description: Human-readable error message
              example: "Invalid email or password"
            request_id:
              type: string
              nullable: true
              description: Request correlation ID for tracing
            details:
              type: object
              nullable: true
              additionalProperties: true
              description: Additional error context
    ValidationErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: integer
                description: Field location path
              msg:
                type: string
                description: Validation error message
              type:
                type: string
                description: Error type identifier
          description: List of field-level validation errors